[#41] BottomViewType 열거형 분리
